requirements.txt
pyrhon-telegram-pot
requests
whisper-openai
pillow
yt-dlp
gtts